🐛 [membership-service] Retry db connection, end program if unsuccessful
When using ArgoCD for deployment, "wait:true" to wait for the database
in the skaffold.yaml, is not applied, instead all services are created
in parallel. The membership-service tries to connect once to the not yet ready db,
and then waits, doing nothing. To mitigate this the option EnableRetryOnFailure
was added. Now the service tries in a timespan of 60 seconds to connect ten times,
and if it still fails exists the program, causing a recreation of the pod.